+++ cpp-breakout/GameStates/MainMenuState.cpp
@@ -11,9 +11,9 @@ MainMenuState::MainMenuState(int levelNumber, StateMachine& machine, sf::RenderW
     : GameState(levelNumber, machine, window, replace) {
 
     // Create texts.
-    title.init(80, sf::Vector2f(m_window.getSize().x / 2, 100), sf::Color(226, 90, 0), "Breakout");
-    play.init(60, sf::Vector2f(m_window.getSize().x / 2, 400), sf::Color(255, 255, 255), "Play");
-    quit.init(60, sf::Vector2f(m_window.getSize().x / 2, 500), sf::Color(255, 255, 255), "Quit");
+    title.init(80, sf::Vector2f((float)m_window.getSize().x / 2, 100), sf::Color(226, 90, 0), "Breakout");
+    play.init(60, sf::Vector2f((float)m_window.getSize().x / 2, 400), sf::Color(255, 255, 255), "Play");
+    quit.init(60, sf::Vector2f((float)m_window.getSize().x / 2, 500), sf::Color(255, 255, 255), "Quit");
 }
 
 void MainMenuState::processEvents() {
