+++ cpp-breakout/Levels/Level.cpp
@@ -26,7 +26,7 @@ void Level::loadLevel(b2World* world, int levelNumber) {
             if(map[i][j].x != -1 && map[i][j].y != -1) {
                 // Go through the numbered grid cells and add the tiles to their appropriate containers.
                 if(map[i][j].x == 0 && map[i][j].y == 0) {
-                    std::unique_ptr<Tile> tempTile(new Tile(world, j * TILE_WIDTH, i * TILE_HEIGHT, 2));
+                    std::unique_ptr<Tile> tempTile(new Tile(world, (float)j * TILE_WIDTH, (float)i * TILE_HEIGHT, 2));
                     tempTile->textRectX = map[i][j].x;
                     tempTile->textRectY = map[i][j].y;
                     tempTile->sprite = tiles;
@@ -38,7 +38,7 @@ void Level::loadLevel(b2World* world, int levelNumber) {
                     solidTiles.push_back(std::move(tempTile));
                 }
                 if(map[i][j].x == 0 && map[i][j].y == 1) {
-                    std::unique_ptr<Tile> tempTile(new Tile(world, j * TILE_WIDTH, i * TILE_HEIGHT, 3));
+                    std::unique_ptr<Tile> tempTile(new Tile(world, (float)j * TILE_WIDTH, (float)i * TILE_HEIGHT, 3));
                     tempTile->textRectX = map[i][j].x;
                     tempTile->textRectY = map[i][j].y;
                     tempTile->sprite = tiles;
@@ -50,7 +50,7 @@ void Level::loadLevel(b2World* world, int levelNumber) {
                     solidTiles.push_back(std::move(tempTile));
                 }
                 if(map[i][j].x == 0 && map[i][j].y == 2) {
-                    std::unique_ptr<Tile> tempTile(new Tile(world, j * TILE_WIDTH, i * TILE_HEIGHT, 6));
+                    std::unique_ptr<Tile> tempTile(new Tile(world, (float)j * TILE_WIDTH, (float)i * TILE_HEIGHT, 6));
                     tempTile->textRectX = map[i][j].x;
                     tempTile->textRectY = map[i][j].y;
                     tempTile->sprite = tiles;
@@ -62,7 +62,7 @@ void Level::loadLevel(b2World* world, int levelNumber) {
                     solidTiles.push_back(std::move(tempTile));
                 }
                 if(map[i][j].x == 0 && map[i][j].y == 3) {
-                    std::unique_ptr<Tile> tempTile(new Tile(world, j * TILE_WIDTH, i * TILE_HEIGHT, 4));
+                    std::unique_ptr<Tile> tempTile(new Tile(world, (float)j * TILE_WIDTH, (float)i * TILE_HEIGHT, 4));
                     tempTile->textRectX = map[i][j].x;
                     tempTile->textRectY = map[i][j].y;
                     tempTile->sprite = tiles;
@@ -74,7 +74,7 @@ void Level::loadLevel(b2World* world, int levelNumber) {
                     solidTiles.push_back(std::move(tempTile));
                 }
                 if(map[i][j].x == 0 && map[i][j].y == 4) {
-                    std::unique_ptr<Tile> tempTile(new Tile(world, j * TILE_WIDTH, i * TILE_HEIGHT, 5));
+                    std::unique_ptr<Tile> tempTile(new Tile(world, (float)j * TILE_WIDTH, (float)i * TILE_HEIGHT, 5));
                     tempTile->textRectX = map[i][j].x;
                     tempTile->textRectY = map[i][j].y;
                     tempTile->sprite = tiles;
@@ -86,7 +86,7 @@ void Level::loadLevel(b2World* world, int levelNumber) {
                     solidTiles.push_back(std::move(tempTile));
                 }
                 if(map[i][j].x == 0 && map[i][j].y == 5) {
-                    std::unique_ptr<Tile> tempTile(new Tile(world, j * TILE_WIDTH, i * TILE_HEIGHT, 1));
+                    std::unique_ptr<Tile> tempTile(new Tile(world, (float)j * TILE_WIDTH, (float)i * TILE_HEIGHT, 1));
                     tempTile->textRectX = map[i][j].x;
                     tempTile->textRectY = map[i][j].y;
                     tempTile->sprite = tiles;
